feat: APIError와 NetworkError를 AppError로 통합.
