+++ HumanNeverDie/Projects/Data/BaseNetwork/Sources/APIRequestable.swift
@@ -34,7 +34,7 @@ public extension APIRequestable {
         }
 
         guard let url = urlComponents?.url else {
-            throw NetworkError.failed(retryable: false, statusCode: -1)
+          throw AppError.network(retryable: false, statusCode: -1)
         }
 
         var request = URLRequest(url: url)
