--- HumanNeverDie/Projects/Data/BaseNetwork/Sources/NetworkError.swift
@@ -1,58 +1,0 @@
-//
-//  NetworkError.swift
-//  BaseNetwork
-//
-//  Created by Seulki Lee on 6/8/25.
-//
-
-import Foundation
-
-public struct APIError: Decodable, Error {
-    public let code: String
-    public let status: Int
-    public let message: String
-    public let path: String?
-    public let timestamp: String?
-
-    public var codeType: APIErrorCode {
-        return APIErrorCode(rawStatus: status)
-    }
-}
-
-public enum NetworkError: Error {
-    case failed(retryable: Bool, statusCode: Int)
-}
-
-public enum APIErrorCode: Int {
-    case badRequest = 400
-    case unauthorized = 401
-    case forbidden = 403
-    case notFound = 404
-    case internalServerError = 500
-    case custom = 600
-    case unknown
-
-    public init(rawStatus: Int) {
-        switch rawStatus {
-        case 400: self = .badRequest
-        case 401: self = .unauthorized
-        case 403: self = .forbidden
-        case 404: self = .notFound
-        case 500: self = .internalServerError
-        case 600...900: self = .custom
-        default: self = .unknown
-        }
-    }
-
-    public var userMessage: String {
-        switch self {
-        case .badRequest: return "요청이 올바르지 않아요."
-        case .unauthorized: return "로그인이 필요해요."
-        case .forbidden: return "권한이 없어요."
-        case .notFound: return "요청한 자원이 없어요."
-        case .internalServerError: return "서버에 문제가 있어요."
-        case .custom: return "처리 중 오류가 발생했어요."
-        case .unknown: return "알 수 없는 에러예요."
-        }
-    }
-}
